Add debug directives to job script
Many debug directives added to job script to trace source of syntax
error when calling python from the job script after loading modules

The source of the error turned out to be the set-alias definition in
the module file.  Fixing this in a custom definition of python-2.7
module allows the job to execute successfully so the error is external
to the job script.